Configuring sass-loader with css-loader & style-loader
Installing & using css-loader with style-loader to watch for css files in the webpack.config. file so it can run the css-loader to read the css file then dump the styles with the style-loader. Also sass is operatiable with the sass-loader which converts all .scss files into regular css so it is transpiled into the styles loader.